old-tests/submission/Microsoft/video/video_022.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'preload' is ignored if 'autoplay' is present.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#attr-media-preload" />
  <meta name="assert" content="'preload' is ignored if 'autoplay' is present." />
</head>
<body>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script type="text/javascript">
    
    
    var test = async_test("'preload' is ignored if 'autoplay' is present.")
    
    function do_loadstart(event)
    {
        test.step(function() { assert_true(true)} );
        test.done();
    }
    
    function timeout()
    {
        test.step(function(){ assert_false(true, "'onloadstart' event never fired!")});
        test.done();
    }
  
    var timer = setTimeout(timeout, 5000);
    
  </script>
  <div id=log></div>
  <video src="broken-uri" preload="none" autoplay onloadstart="do_loadstart();">Your browser does not support media elements.</video>
</body>
</html>
